fixed path in prepareSolarForecast
