+++ portal/app/algorithms/SARIMA_SolarModel_params960.npz
